requirements.txt
python-dotenv
autogen
google-genai
google-cloud-aiplatform
ag2[gemini]